test(streamlit): write unit tests for imednet-streamlit plugin (#1008)
